app/requirements.txt

fastapi==0.116.1
passlib==1.7.4
pydantic==2.11.7
PyJWT==2.10.1
python-dotenv==1.1.1
SQLAlchemy==2.0.36
bcrypt==3.2.2
apscheduler==3.10.4
aiohttp==3.11.11
twilio==9.2.3
